requirements.txt
discord.py
Pillow
websocket-client
ollama
wordcloud
nltk
